+++ packages/Documents/src/components/com_documents/domains/entities/component.php
@@ -10,4 +10,27 @@
  *
  * @link       http://www.GetAnahita.com
  */
-class ComDocumentsDomainEntityComponent extends ComMediumDomainEntityComponent{}
+class ComDocumentsDomainEntityComponent extends ComMediumDomainEntityComponent
+{
+    /**
+     * {@inheritdoc}
+     */
+    protected function _setGadgets($actor, $gadgets, $mode)
+    {
+        if ($mode == 'profile') {
+            $gadgets->insert('documents', array(
+                    'title' => AnTranslator::_('COM-DOCUMENTS-GADGET-ACTOR-PROFILE'),
+                    'url' => 'option=com_documents&view=documents&layout=gadget&oid='.$actor->id,
+                    'action' => AnTranslator::_('LIB-AN-GADGET-VIEW-ALL'),
+                    'action_url' => 'option=com_documents&view=documents&oid='.$actor->id,
+            ));
+        } else {
+            $gadgets->insert('documents', array(
+                    'title' => AnTranslator::_('COM-DOCUMENTS-GADGET-ACTOR-DASHBOARD'),
+                    'url' => 'option=com_documents&view=documents&layout=gadget&filter=leaders',
+                    'action' => AnTranslator::_('LIB-AN-GADGET-VIEW-ALL'),
+                    'action_url' => 'option=com_documents&view=documents&filter=leaders',
+            ));
+        }
+    }
+}